app/requirements-dev.txt
django-debug-toolbar==3.2.1
django-extensions==3.1.0
